Extract more information
The original output was simple and working for the most trivial cases.
There exist a number of other cases for which the existing output was
confusing: reducing or silencing output that makes no sense as well as
querying more keys and metadata makes things a bit more clear although
encrypted content is not decrypted and the key hierarchy isn't rebuilt.

Also, the public key wasn't hashed as required before being converted to
Base58Check, and compressed keys are handled somehow (hack-ish).